Modification to upgrades combined with changes to the ui

+++ core/src/main/java/com/project/clicker/logic/PopulationManager.java
@@ -7,7 +7,7 @@ public class PopulationManager {
         this.state = state;
     }
 
-    public void increasePopulation(int people) {
+    public void increasePopulation(long people) {
         state.addPopulation(people);
     }
